DeepUMQA-Global/requirements.txt
# Minimal runtime for inference with precomputed full-length features.
# Keep NumPy on the 1.x line unless you rebuild the PyTorch / PyG binary stack.
numpy>=1.21,<2
scipy>=1.7,<2
biopython>=1.79
tqdm>=4.64
pandas>=1.3

# Install matching CPU/CUDA builds for your platform.
torch>=1.11,<2.1
torch-geometric>=2.0,<2.4

# PyG compiled extensions required by the current model code.
torch-scatter>=2.0,<2.2
torch-sparse>=0.6,<0.7
torch-cluster>=1.6,<1.7
torch-spline-conv>=1.2,<1.3
